add datavariant tests

See merge request 373-22-3-injera/social-gaming!57

+++ lib/game/include/ContentVariant.h
@@ -49,7 +49,7 @@ struct toFloatVisitor{
 	float operator()(float f) const;
 
 	template <typename T>
-    int operator()(T) const{
+    float operator()(T) const{
     throw std::invalid_argument{"invalid"};
 	}
 };
@@ -62,7 +62,6 @@ struct shuffleVisitor {
     throw std::invalid_argument{"invalid"};
 	}
 };
-
 
 struct addVisitor
 {
